LV-CIT-main/LV-CIT-main/requirements.txt
descartes==1.1.0
imutils==0.5.4
inplace-abn==1.1.0  # NOTE: for Windows users: please install microsoft visual studio and add PATH_TO_BIN (for MSVC) to system environment before installing inplace_abn; for Linux users: please download the source code from github (https://github.com/mapillary/inplace_abn/releases) and install inplace_abn from source code.
matplotlib==3.3.4
numpy==1.24.4
opencv-python==4.4.0.46
pandas==1.3.0
Pillow==10.3.0
prefetch-generator==1.0.1
pycocotools==2.0.6
scikit-learn==0.24.2
Shapely==1.8.5.post1
tqdm==4.66.3
boto3==1.34.5
torchnet==0.0.4
openpyxl==3.0.10
pandarallel==1.6.5
